Feature/p5 notification

+++ app/src/main/java/com/and/apartmentmanager/data/local/entity/InviteCodeEntity.java
@@ -45,5 +45,5 @@ public class InviteCodeEntity {
     private boolean isUsed;
 
     @ColumnInfo(name = "used_by")
-    private Integer usedBy;        // nullable
+    private Long usedBy;        // nullable
 }
